Rename api to routes
